pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repostats"
version = "0.1.0"
description = "A CLI tool to fetch GitHub repository statistics"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = ["github", "statistics", "cli", "api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Build Tools",
]
dependencies = [
    "requests>=2.28.2",
    "click>=8.1.3",
    "PyYAML>=6.0.1",
]

[project.urls]
Homepage = "https://github.com/sahansera/repostats"
Repository = "https://github.com/sahansera/repostats"
Issues = "https://github.com/sahansera/repostats/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "pytest-cov>=4.1.0",
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.3.0",
    "types-requests>=2.32.0",
]

[project.scripts]
repostats = "cli:main"

[tool.setuptools]
py-modules = ["__init__", "cli", "github"]
package-dir = {"" = "src"}

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
